# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API",
    "version": "1.0.0",
    "description": "API documentation for my Express application"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "get": {
        "summary": "Render the login page",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Render the login form"
          }
        }
      },
      "post": {
        "summary": "Log in an existing user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username of the user"
                  },
                  "password": {
                    "type": "string",
                    "description": "The password of the user"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Login successful, redirect to user role dashboard"
          },
          "401": {
            "description": "Invalid username or password"
          }
        }
      }
    },
...
    "/courses": {
      "get": {
        "summary": "Retrieve a list of all courses",
        "responses": {
          "200": {
            "description": "A list of courses"
          },
          "401": {
            "description": "Unauthorized access"
          }
        }
      }
    },
...
    "/student/courses/enroll/{id}": {
      "get": {
        "summary": "Enroll in a course",
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the course to enroll in",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully enrolled in the course"
          }
        }
      }
    },
...
    "/quizzes/{id}/enroll": {
      "post": {
        "summary": "Enroll in a quiz",
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the quiz to enroll in",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully enrolled in the quiz"
          }
        }
      }
    },
...
    "/assignments/{assignmentId}/submit": {
      "post": {
        "summary": "Submit an assignment",
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "description": "The ID of the assignment to submit",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assignment submitted successfully"
          }
        }
      }
    },
...
    "/student/grades": {
      "get": {
        "summary": "View student's grades",
        "tags": [
          "Students"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved student's grades",
            "content": {
              "text/html": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "assignmentGrades": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "assignmentId": {
                            "type": "string",
                            "description": "The ID of the assignment"
                          },
                          "title": {
                            "type": "string",
                            "description": "The title of the assignment"
                          },
                          "grade": {
                            "type": "string",
                            "description": "The grade achieved in the assignment"
                          }
                        }
                      }
                    },
                    "quizGrades": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "quizId": {
                            "type": "string",
                            "description": "The ID of the quiz"
                          },
                          "title": {
                            "type": "string",
                            "description": "The title of the quiz"
                          },
                          "grade": {
                            "type": "string",
                            "description": "The grade achieved in the quiz"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "description": "A message indicating the status of the grades"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "An error occurred while fetching grades"
          }
        }
      }
    },
...
    "/respond/{notificationId}": {
      "post": {
        "summary": "Respond to a notification",
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "description": "The ID of the notification to respond to",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Response data for the notification",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "response": {
                    "type": "string",
                    "description": "The student's response"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response submitted successfully"
          }
        }
      }
    },
...
  },
...
}